+++ packages/hms_room_kit/lib/src/hls_viewer/hls_player_overlay_options.dart
@@ -6,6 +6,7 @@ import 'package:flutter/material.dart';
 ///Project imports
 import 'package:hms_room_kit/src/hls_viewer/hls_viewer_bottom_navigation_bar.dart';
 import 'package:hms_room_kit/src/hls_viewer/hls_viewer_header.dart';
+import 'package:hms_room_kit/src/hls_viewer/hls_viewer_mid_section.dart';
 
 ///[HLSPlayerOverlayOptions] renders the overlay options for the HLS Player
 class HLSPlayerOverlayOptions extends StatelessWidget {
@@ -14,16 +15,30 @@ class HLSPlayerOverlayOptions extends StatelessWidget {
   const HLSPlayerOverlayOptions({super.key, required this.hasHLSStarted});
   @override
   Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+    ///Here top and bottom navigation bar are in Column
+    ///while mid section is above the Column
+    ///This is done to avoid overflowing in case of
+    ///large transcription
+    return Stack(
       children: [
-        HLSViewerHeader(
-          hasHLSStarted: hasHLSStarted,
+        Column(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            HLSViewerHeader(
+              hasHLSStarted: hasHLSStarted,
+            ),
+
+            ///Renders the bottom navigation bar if the HLS has started
+            ///Otherwise does not render the bottom navigation bar
+            hasHLSStarted ? HLSViewerBottomNavigationBar() : const SizedBox()
+          ],
         ),
 
-        ///Renders the bottom navigation bar if the HLS has started
-        ///Otherwise does not render the bottom navigation bar
-        hasHLSStarted ? HLSViewerBottomNavigationBar() : const SizedBox()
+        ///This renders the pause/play button
+        Align(
+          alignment: Alignment.center,
+          child: hasHLSStarted ? HLSViewerMidSection() : const SizedBox(),
+        )
       ],
     );
   }