+++ packages/hmssdk_flutter/android/src/main/kotlin/live/hms/hmssdk_flutter/hls_player/HMSHLSPlayerAction.kt
@@ -31,6 +31,7 @@ class HMSHLSPlayerAction {
                 "are_closed_captions_supported" -> areClosedCaptionsSupported(result)
                 "enable_closed_captions" -> enableClosedCaptions(result)
                 "disable_closed_captions" -> disableClosedCaptions(result)
+                "get_stream_properties" -> getStreamProperties(result)
                 else -> {
                     result.notImplemented()
                 }
@@ -54,7 +55,7 @@ class HMSHLSPlayerAction {
             val hlsUrl = call.argument<String?>("hls_url")
             hlsActions?.let {
                 it.get()?.start(hlsUrl, result)
-            }?:run{
+            } ?: run {
                 HMSErrorLogger.logError("start", "hlsActions is NULL", "NULL Error")
             }
         }
@@ -67,7 +68,7 @@ class HMSHLSPlayerAction {
         private fun stop(result: Result) {
             hlsActions?.let {
                 it.get()?.stop(result)
-            }?:run{
+            } ?: run {
                 HMSErrorLogger.logError("stop", "hlsActions is NULL", "NULL Error")
             }
         }
@@ -80,7 +81,7 @@ class HMSHLSPlayerAction {
         private fun pause(result: Result) {
             hlsActions?.let {
                 it.get()?.pause(result)
-            }?:run{
+            } ?: run {
                 HMSErrorLogger.logError("pause", "hlsActions is NULL", "NULL Error")
             }
         }
@@ -93,7 +94,7 @@ class HMSHLSPlayerAction {
         private fun resume(result: Result) {
             hlsActions?.let {
                 it.get()?.resume(result)
-            }?:run{
+            } ?: run {
                 HMSErrorLogger.logError("resume", "hlsActions is NULL", "NULL Error")
             }
         }
@@ -106,7 +107,7 @@ class HMSHLSPlayerAction {
         private fun seekToLivePosition(result: Result) {
             hlsActions?.let {
                 it.get()?.seekToLivePosition(result)
-            }?:run{
+            } ?: run {
                 HMSErrorLogger.logError("seekToLivePosition", "hlsActions is NULL", "NULL Error")
             }
         }
@@ -128,9 +129,9 @@ class HMSHLSPlayerAction {
                 }
 
             seconds?.let {
-                hlsActions?.let {_hlsActions ->
+                hlsActions?.let { _hlsActions ->
                     _hlsActions.get()?.seekForward(it, result)
-                }?:run{
+                } ?: run {
                     HMSErrorLogger.logError("seekForward", "hlsActions is NULL", "NULL Error")
                 }
             }
@@ -155,7 +156,7 @@ class HMSHLSPlayerAction {
             seconds?.let {
                 hlsActions?.let { _hlsActions ->
                     _hlsActions.get()?.seekBackward(it, result)
-                }?:run{
+                } ?: run {
                     HMSErrorLogger.logError("seekBackward", "hlsActions is NULL", "NULL Error")
                 }
             }
@@ -178,9 +179,9 @@ class HMSHLSPlayerAction {
                 }
 
             volume?.let {
-                hlsActions?.let {_hlsActions ->
-                    _hlsActions.get()?.setVolume(it,result)
-                }?:run{
+                hlsActions?.let { _hlsActions ->
+                    _hlsActions.get()?.setVolume(it, result)
+                } ?: run {
                     HMSErrorLogger.logError("setVolume", "hlsActions is NULL", "NULL Error")
                 }
             }
@@ -192,10 +193,9 @@ class HMSHLSPlayerAction {
          * @param result The result object to be returned after adding the HLS stats listener.
          */
         private fun addHLSStatsListener(result: Result) {
-
             hlsActions?.let {
                 it.get()?.addHLSStatsListener(result)
-            }?:run{
+            } ?: run {
                 HMSErrorLogger.logError("addHLSStatsListener", "hlsActions is NULL", "NULL Error")
             }
         }
@@ -208,7 +208,7 @@ class HMSHLSPlayerAction {
         private fun removeHLSStatsListener(result: Result) {
             hlsActions?.let {
                 it.get()?.removeHLSStatsListener(result)
-            }?:run{
+            } ?: run {
                 HMSErrorLogger.logError("removeHLSStatsListener", "hlsActions is NULL", "NULL Error")
             }
         }
@@ -222,7 +222,7 @@ class HMSHLSPlayerAction {
         private fun areClosedCaptionsSupported(result: Result) {
             hlsActions?.let {
                 it.get()?.areClosedCaptionsSupported(result)
-            }?:run{
+            } ?: run {
                 HMSErrorLogger.logError("areClosedCaptionsSupported", "hlsActions is NULL", "NULL Error")
             }
         }
@@ -235,7 +235,7 @@ class HMSHLSPlayerAction {
         private fun enableClosedCaptions(result: Result) {
             hlsActions?.let {
                 it.get()?.enableClosedCaptions(result)
-            }?:run{
+            } ?: run {
                 HMSErrorLogger.logError("enableClosedCaptions", "hlsActions is NULL", "NULL Error")
             }
         }
@@ -248,8 +248,16 @@ class HMSHLSPlayerAction {
         private fun disableClosedCaptions(result: Result) {
             hlsActions?.let {
                 it.get()?.disableClosedCaptions(result)
-            }?:run{
+            } ?: run {
                 HMSErrorLogger.logError("disableClosedCaptions", "hlsActions is NULL", "NULL Error")
+            }
+        }
+
+        private fun getStreamProperties(result:Result){
+            hlsActions?.let {
+                it.get()?.getStreamProperties(result)
+            } ?: run {
+                HMSErrorLogger.logError("getStreamProperties", "hlsActions is NULL", "NULL Error")
             }
         }
     }