+++ packages/hmssdk_flutter/android/src/main/kotlin/live/hms/hmssdk_flutter/hls_player/IHLSPlayerActionInterface.kt
@@ -45,4 +45,6 @@ interface IHLSPlayerActionInterface {
     fun enableClosedCaptions(result: Result)
 
     fun disableClosedCaptions(result: Result)
+
+    fun getStreamProperties(result: Result)
 }